requirements.txt
asgiref==3.5.0
Django==3.2.7
django-filter==21.1
django-model-utils==4.1.1
djangorestframework==3.12.4
Pillow==9.0.0
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
djangorestframework-simplejwt
celery
redis
django-environ
drf-yasg==1.20.0
scipy
kombu==4.2.1